% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboheatmap.R
\name{riboheatmap}
\alias{riboheatmap}
\title{A heatmap Function}
\usage{
riboheatmap(gene_list, cl = FALSE)
}
\arguments{
\item{cl}{Do you analyze by cell line? Defaults to FALSE}
}
\description{
This function allows you to express your love of cats.
}
\examples{
riboheatmap()
}
\keyword{heatmap}
